refactor: Support fractional travel hours (double instead of int)
Changed TravelCostHours from int? to double? to allow sub-hour travel (0.5 hours).
Updated all related handlers and models:
- LocationExit.TravelCostHours now accepts double
- TravelChange.TravelCostHoursOverride now accepts double
- EncounterResolver.EvaluateAsync accepts double totalHours
- CampaignTime.AdvanceHours rounds double to nearest whole hour

Also enhanced guidance for Conversation event requirements in ToolCallExamples
and CommitEnumCheatSheet to clarify that involved field is REQUIRED for
Conversation events but optional for others. Updated recommended-system-prompt.md.

All 1134 unit tests pass.

+++ src/CampaignVault/Data/ChangeHandlers/TravelChangeHandler.cs
@@ -89,7 +89,7 @@ public class TravelChangeHandler : IWorldChangeHandler
         {
             time.AdvanceHours(hoursTraveled);
 
-            var tirednessDelta = (hoursTraveled / 4.0f) * 10f;
+            var tirednessDelta = (float)((hoursTraveled / 4.0) * 10.0);
             if (tirednessDelta > 0)
             {
                 await context.Dispatcher.DispatchMutationAsync(context, new NeedChange
